Limit order amount
Fix #85
Move out check_env_vars() to utils
