+++ .env-sample
@@ -30,3 +30,5 @@ EXP_SECONDS=900
 MAX_ROUTING_FEE=0.001
 # Mostro Fee
 FEE=0.006
+# Max order amount (sats)
+MAX_ORDER_AMOUNT=10000
