+++ src/util.rs
@@ -442,3 +442,54 @@ pub async fn settle_seller_hold_invoice(
 
     Ok(())
 }
+
+pub fn check_env_vars() -> Result<()> {
+    // Mandatory env vars
+    let _ = var("NSEC_PRIVKEY")
+        .context("Missing NSEC_PRIVKEY env variable from env file - add mostro private key")?;
+    let _ = var("RELAYS")
+        .context("Missing RELAYS env variable from env file - add relay list comma separated")?;
+    let _ = var("DATABASE_URL").context("Missing DATABASE_URL from env file - Add a path")?;
+    let _ = var("LND_CERT_FILE").context("Missing LND_CERT_FILE from env file - Add a path")?;
+    let _ =
+        var("LND_MACAROON_FILE").context("Missing LND_MACAROON_FILE from env file - Add a path")?;
+    let _ = var("LND_GRPC_PORT")
+        .context("Missing LND_GRPC_PORT from env file - Add port value")?
+        .parse::<u64>()
+        .context("Error parsing LND_GRPC_PORT")?;
+    let _ = var("LND_GRPC_HOST").context("Missing LND_GRPC_HOST from env file - set host value")?;
+    let _ = var("INVOICE_EXPIRATION_WINDOW")
+        .context("Missing INVOICE_EXPIRATION_WINDOW from env file - Add expiration value")?
+        .parse::<u64>()
+        .context("Error parsing INVOICE_EXPIRATION_WINDOW")?;
+    let _ = var("HOLD_INVOICE_CLTV_DELTA")
+        .context("Missing HOLD_INVOICE_CLTV_DELTA from env file - Add cltv invoice value")?
+        .parse::<u64>()
+        .context("Error parsing HOLD_INVOICE_CLTV_DELTA")?;
+    let _ = var("MIN_PAYMENT_AMT")
+        .context("Missing MIN_PAYMENT_AMT from env file - Add min payment value")?
+        .parse::<u64>()
+        .context("Error parsing MIN_PAYMENT_AMT")?;
+    let _ = var("EXP_SECONDS")
+        .context("Missing EXP_SECONDS from env file - Add expiration invoice seconds value")?
+        .parse::<u64>()
+        .context("Error parsing EXP_SECONDS")?;
+    let _ = var("EXP_HOURS")
+        .context("Missing EXP_HOURS from env file - Add expiration order hours value")?
+        .parse::<u64>()
+        .context("Error parsing EXP_HOURS")?;
+    let _ = var("MAX_ROUTING_FEE")
+        .context("Missing MAX_ROUTING_FEE from env file - Add routing fees value")?
+        .parse::<f64>()
+        .context("Error parsing MAX_ROUTING_FEE")?;
+    let _ = var("FEE")
+        .context("Missing FEE from env file - Add mostro fees value")?
+        .parse::<f64>()
+        .context("Error parsing FEE")?;
+    let _ = var("MAX_ORDER_AMOUNT")
+        .context("Missing MAX_ORDER_AMOUNT from env file - Limit publish orders to a max amount")?
+        .parse::<f64>()
+        .context("Error parsing MAX_ORDER_AMOUNT")?;
+
+    Ok(())
+}
